requirements.txt
PyNaCl~=1.4.0
